psg: Use float for sample_counter
To make sure that we don't lose samples since PSG_CLOCKS_PER_SAMPLE is not
integer.

Still sounds like shit.

+++ megado/psg.h
@@ -28,7 +28,7 @@ typedef struct NoiseChannel {
 
 typedef struct PSG {
     uint16_t remaining_clocks;
-    uint16_t sample_counter;
+    float sample_counter;
 
     SquareChannel square[3];
     NoiseChannel noise;
